Add Ada project file and initial source files.
Commit the first version of my Ada translation of the Apple II
memory management code and all of the associated data types
and structs. I've tried to stay fairly close to the original
AppleWin / LinApple C++ code in structure. There's also a
GNAT GPRbuild project file (apple2.gpr) and Alire package
metadata file (alire.toml).

The main procedure is a placeholder that makes a few calls to
exercise the functionality that I've implemented so far. The
important thing is that it runs without throwing any constraint
errors or assertion failures. That was a big milestone for me!

My hope is to try to validate as much of the emulator code as
possible as SPARK 2014 (safety-critical variant of Ada 2012),
so I've simplified the memory map to use a single hard-coded
slot configuration that matches the default LinApple config:

 - Slot 1: Super Serial Card
 - Slot 2: parallel printer interface
 - Slot 4: Phasor / Mockingboard sound
 - Slot 5: (same as slot 4)
 - Slot 6: Disk ][ card

You will be able to select from any of the four model variants,
Apple II, II Plus, IIe, and IIe Enhanced, as in the original.

My hope is to use SDL 2 with the SDLAda bindings on GitHub to
support video, keyboard, joystick, speaker, Mockingboard audio,
one or two Disk ][ disk drives, and RAMWorks III 8 MB RAM card.

I don't plan to support the mouse, clock, or hard drive devices
initially, since they're only available in the Linux version
by editing the code or hand-editing the "registry" config file.

Downloading disk images via FTP and the debugger are also not
going to be a high priority for me, but I should have everything
else converted from C++ to Ada for version 1.0. Unlike LinApple,
this version should build on Windows as well as Linux and macOS.
I'm using some GNAT-specific pragmas, but nothing OS-specific.