Fix bugs and refactor for SPARK compliance.
Refactored the new code to use a tagged record type for the Computer,
deriving from the new MOS_CPU_6502.CPU_6502_Series base class, with two
abstract procedures for the CPU emulation to use to read and write to
the emulated address space. All of the emulation procedures now take a
Computer object, or one of its parent classes, as a parameter.

SPARK has strict rules about when classes must be "frozen", which
led me to end up making an Apple2_Base class which doesn't implement
the abstract read and write procedures, plus a Computer subclass in
Apple2.Memory that inherits from Apple2_Base and implements Mem_IO_Read
and Mem_IO_Write. The subclass requires "pragma Elaborate_Body;" in
its package spec, which means none of the other packages can call
methods on Computer directly without a circular build dependency.
Adding the middle Apple2_Base class breaks the circular dependency.

The code now passes GNATprove static analysis in Bronze (flow analysis)
and Silver (proof) modes. To achieve this, I removed nearly all of the
pointers, creating a single 8 MB RAM + ROM array consisting of the main
64K RAM in bank 0, internal ROM in bank 1, peripheral ROMs in bank 2,
and the 128 banks supported by RAMWorks III in banks 3 to 130. Bank 3 is
also the second 64K of a 128K Apple IIe configuration. The highest
bank of RAM accessed is recorded so that snapshots won't be larger than
needed. All reads and writes use array index calculations, not pointers.

Unfortunately, the RAM / ROM bank byte array can't be contained within
the Apple2_Base or Computer records because when I attempted to do so,
GNAT immediately fails with a runtime constraint error at the point of
attempting to allocate the variable or pass it as a parameter. If I try
to allocate it dynamically, I immediately run into SPARK restrictions
and confusion about how to dynamically dispatch on access to an object.

I've had to break the encapsulation and make "Mem" a separate parameter
passed by access along with the Computer object to the procedures that
require either read or write access to the RAM / ROM. This includes
expected uses in Disk I/O and Mockingboard audio, and expected non-uses
of memory by joystick, keyboard, printer, speaker, and serial devices.

This decision had the unexpected side benefit of making explicit which
parts of the emulation access the emulated memory and which don't. One
perhaps surprising detail is that many reads in $Cxxx that don't
explicitly connect to I/O or ROM return a "floating bus" value which
represents the memory that the video update logic had been reading from
RAM at the time of the CPU read, possibly with the high bit of the byte
modified. This is why there's a number of cycles elapsed parameter in
many of the $Cxxx region procedures and associated device I/O routines.

Because there's a single instance of the memory created in my test main
procedure, the access to it doesn't have to be "access all" and
therefore Ada and SPARK appear to be relatively happy with passing
access to the type because it can statically prove the ownership.
I do have to allocate it on the heap with "new", rather than declaring
the 8MB array as a local stack variable (another Constraint_Error).

As another concession to SPARK rules, the Mem_Read_Random_Data function
that will be used by the Disk II emulation is now in an Apple2.Random
package with its body defined with SPARK_Mode => Off, because the system
random number package is marked SPARK_Mode => Off due to being defined
as a function with the side effect of updating the Generator parameter
(SPARK functions are supposed to be free of side effects). Since SPARK
mode is disabled inside Apple2.Random's body (but not the spec), it
shouldn't be bothered by the package private global Generator variable.