Simplify project file for use with GNAT Studio.
My original .gpr file was based on the GPRbuild file for building
gprbuild and the other executables in the GPRbuild project. The
definition of build type and the number of build processes details,
along with separate compiler flags for different modes, is redundant
with the functionality of GNAT Studio for adding "-O2", "-g", etc. on
the GPRbuild commandline and automatically creating appropriate
subdirectories in the specified object and executables directories.

GNAT Studio does recognize and allow you to specify what it calls
"scenario variables" for the project, but it also has its own build mode
menu of choices that duplicates the local ones in the project, and adds
additional build types. It made the most sense to me to remove the
redundant custom build type and "-j" build parallel options.