Minor changes to clarify code.
Moved code comment lines inside procedures above the code they're
commenting on. Renamed the "Jammed" Boolean of CPU_6502_Series to
Halt_Opcode and changed the type to Unsigned_8 so it can contain the
illegal opcode that halted the NMOS 6502, for display purposes. Since
opcode $00 is BRK, which is handled, there should be no ambiguity as to
the mode when this is non-zero.

I also optimized the memory reset code that fills the main RAM with a
pattern of $FF and $00 (by default) to fill all four bytes instead of
first zero-filling the entire area and then writing the two non-zero
bytes and skipping two. This should be just a tiny bit faster. The
two 64K regions are still zero-filled for the non-pattern-fill case.