Remove write page dirty bits table.
As I prepare to start porting the video emulation code, and based on my
past experience starting to write an Amiga 4000 machine type for QEMU,
it makes the most sense to me for the CPU emulation to call the video
emulation once per scanline, so it can see a view of RAM that is very
close to the CPU's view, hopefully without too much cache thrashing.

I want to support as many clever video tricks as were used in Apple IIe
software, and I don't think it makes sense to try to get the horizontal
pixel clock interleaved with the CPU accessing one byte of RAM, as on
the real machine. I think that scanline-level interleaving should be
fine, as long as it has the ratio of 7 or 14 bits (1 or 2 bytes) of
reads from video RAM per cycle. The way the display works is like an
extremely bare-bones and simplified version of the Amiga chipset.

Because I'm going to write the video this way, there's no need for every
memory write to update a dirty page table to determine whether or not to
update a frame. The display is always going to be updated every frame.
Therefore, I'm removing that code in this commit.