Refactor memory and I/O emulation and CPU API.
Refactor the memory management code for readability and reliability.
I've removed the 256-byte read and write page table arrays, merged most
"Read" and "Write" procedures into a single "Access" (which is how the
original C++ code was structured), and changed Mem_IO_Access to compute
the bank to read/write on every bus access, instead of looking it up.

The new Mem_IO_Access tests for zero/stack page first (controlled by
the ALTZP switch), then main RAM ($0200 .. $BFFF), which is controlled
by multiple switches, then bank-switched RAM/ROM ($D000 .. $FFFF), also
controlled by multiple switches, including a write-enable switch, then
finally the 4K I/O / ROM space at $C000 .. $CFFF.

The $Cxxx region is complex enough to require several helper functions,
which I've refactored into a smaller number of hopefully more readable
versions. Mem_IO_Access calls IO_Access_Cxxx, which in turn looks at the
middle two nybbles of the address to determine which device emulation or
soft switch handler to use. Writes to $C100 .. $CFFF are ignored, except
where a device in a slot is using their $Csxx space for I/O. Currently,
this only applies to the Mockingboard sound card, assigned to slots 4
and 5 in the default LinApple configuration. Once I add mouse emulation
as an option for slot 4, enabling it will remove the sound card there.

Reads from $C100 .. $CFFF outside of slot 4/5 space go to IO_Read_C1xx,
which determines whether to return internal ROM, peripheral ROM, or
floating bus (emulating video read / refresh of DRAM) based on several
switches. Now that I'm not using a page table for lookups, I don't need
two 64K banks to hold ROM data in order to read from it, so the ROMs no
longer need to be copied, freeing 128K of mostly wasted address space
that was in my original design.

I wanted to use access variables for all ROMs, but Ada requires array
ranges to match exactly and won't allow 'Access on array slices (GNAT
complains that the slices aren't aliased, even if the full array is).
It also won't do array sliding of an access type like Ada does for
copying ranges, so there's an additional error when the start isn't 0.

The Apple IIe (Enhanced) ROMs need to be split into 12K main and 4K
extra pieces, and the SSC ROM needs a 256-byte slice of its 2K expansion
ROM for its 256-byte ROM area. It was easiest to copy the ROMs into the
Computer object by value, which saves a few null checks, at least. For
the 2K expansion ROMs, there's only SSC at the moment (the mouse card
will add another), and they don't need any slicing, so I used an array
of access variables for those ROMs without any problems.

The CPU emulation is now going to increment vertical and horizontal
counters on each clock cycle to make it easy to read the floating bus
value when a program accesses an unmapped I/O location, as well as for
efficient emulation of one frame at a time (start VBL to start VBL).
The CPU emulation will update the counters in the CPU_6502_Series
object before any calls to Mem_IO_Access, eliminating the need to pass
cycle counts into every device handler like the original code does.

Added debug lines to the stub implementations of the devices, as well as
to note accesses to set the annunciators, which could be the result of
code trying to set double hi-res mode with the DHIRES switch hidden, or,
for Apple II/II+ emulation, a program thinking it's running on a IIe.

I'm going to convert the 6502 / 65C02 emulation next, to validate my
plan to emulate one or more scan lines efficiently (by passing the
current column and scan line in and out, along with the number of
columns (so the CPU knows when to go to the next scan line) and the
number of scan lines to emulate (either 1, or 70/120 lines in VBL).

When the display is visible, the video and CPU emulation will alternate
for each scan line, so any code "chasing the beam" will have a better
chance of working than if I rendered the video on a per-frame basis.
The SDL events and associated emulated state will be refreshed in
between frames, so by starting the frame emulation at the start of VBL,
I hope to minimize latency by giving emulated code the entire VBL and
the following frame to poll and get the current user input state.

Removed the lines to patch the Disk II ROM to remove a timing loop. My
plan is for the disk emulation to be accurate and run at normal speed.