Add new Ada conversion of 65C02 emulation.
Add new Ada conversion of 65C02/6502 emulation from LinApple/AppleWin.
The new version should correctly calculate the extra cycles from page
crossing (the original sets "bSlowerOnPagecross" in the opcode macros
after it's tested in the addressing mode macros, rather than before).

The Ada version should be a bit easier to understand and reason about.
GNATprove's data flow analysis found a bug in one line where I hadn't
passed enough parameters, because it noticed that a value was being set
and not used. I'm not sure how to suppress the warning that PC is being
updated and then not used for the two JMP instructions that use indexed
addressing modes and then update PC with the target address (for the
absolute JMP, I inlined the memory read, but it'd be less readable to do
that for the other two cases).

This version implements the new Rockwell/WDC 65C02 instructions BBR,
BBS, RMB, and SMB in the $x7 and $xf opcodes (32 opcodes in all, one for
each of the bits you can set or clear), as well as the behavior of the
one-cycle NOP and the other NOPs.

I've converted Cycles_Since_Boot from a variable to a function that
returns the value by multiplying frames since boot by the additional
cycles from the start of the frame. The scan line and column counter are
the only variables being updated within a frame, so code that needs to
compute a duration needs the exact cycle count, and this feels optimal,
especially since the calling code didn't have to change.

I've added a benchmark based on the LinApple version, but emulating one
scanline at a time during the visible frame instead of 100,000 cycles at
a time, to verify my plan to alternate between CPU and video emulation.
I'm seeing ~140 MHz with either CPU core, in NTSC or PAL, on my Xeon.
I'm going to look at optimizing the memory and I/O bus emulation, and
experimenting to see how much CPU state to copy into local variables.

I ended up using all unsigned types in the CPU emulation and
sign-extending the 8-bit branch displacements to 16 bits, since it
turned out to be too convoluted to convert between signed and unsigned
integer types without raising Constraint_Error, e.g. the negation of
-128 doesn't fit in a signed byte, and you can't use bitwise "and" with
signed types. I could probably use Ada.Unchecked_Conversion, but didn't
want to cause problems for GNATprove and SPARK_Mode.