Inline memory reads/writes using per-page tables.
Optimize RAM reads/writes by using per-256-byte-page lookup tables
similar to my previous design, and the LinApple code, except now
using 16-bit bank and page offsets instead of an 8-bit bank entry, or
byte pointers, as LinApple uses. Now the benchmark of frequently-used
opcodes runs at ~670 MHz on my Xeon, instead of ~140 MHz, and there are
no obvious hotspots to fix in the perf report.

Now, changes to the memory map soft switches result in a call to
Mem_Update_Paging, which always updates all 256 read and write entries.
For the $D000 double mapping, the $1000 offset to access the 4K RAM at
$Cxxx in each memory bank is now subtracted from the page that's now
part of the lookup, instead of a special case.

When high RAM ($D000-$FFFF) isn't writable, the write page table entries
will all point to the first 256 bytes of the ROM bank ($0100xx), where
writes won't overwrite anything and can't be read. This optimizes the
now-inlined default path for reads and writes to a single lookup, bit
shift, mask and or with the low byte of the address to access. The high
byte of the address is now incorporated into the value in the table, and
cases like the two just described are easily handled.

Now, the only special case for the Apple II series is the 4K at $Cxxx,
which needs special handling for writes and even for reads from ROM,
because those reads can trigger different ROM options to be swapped in
in the upper 2K at $C800-$CFFF. Those cases are handled, as previously,
by overridden procedures, now named Mem_IO_Read_Special and
Mem_IO_Write_Special (and no longer merged into a single "Access"),
which are called only when the page table lookup returns 16#FFFF#.