Add comment about CPU_6502_Series register sizes.
The program bank and data bank registers that I added for future 65C816
emulation are actually 8-bit registers, but I discovered that the 65C02
emulation, which doesn't use them, ran faster with 16-bit variables.
Later testing returned faster results with 8-bit variable types.

I'm thankful for the AppleWin / LinApple CPU benchmark with its
groupings of common opcodes because I never would've suspected that
the layout of the CPU registers in the record would have such an impact.
The placement of Halt_Opcode is similarly based on experimentation.